requirements.txt
fastapi>=0.111,<1.0
uvicorn>=0.29,<0.40
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
httpx>=0.27,<0.30
websockets>=12,<15
nats-py>=2.7,<3
redis>=5,<6
SQLAlchemy>=2.0,<3
asyncpg>=0.29,<1
python-dotenv>=1,<2
prometheus-client>=0.20,<1
loguru>=0.7,<1
tenacity>=8.2,<10
optuna>=3.6,<5
